This reverts 8f7f28a7c3db28711920ad5d15df401fae90bcf1 (OpenGL: Upgrade
to
Mesa 9.2).

Conflicts:
	build/jam/BuildFeatures
	build/jam/HaikuImage
	build/jam/OptionalPackages
	build/scripts/build_cross_tools_gcc4
	src/add-ons/opengl/swpipe/Jamfile
	src/apps/diskusage/Jamfile
	src/kits/tracker/ContainerWindow.cpp
	src/kits/tracker/DeskWindow.cpp
	src/kits/tracker/Jamfile

+++ src/apps/processcontroller/KernelMemoryBarMenuItem.cpp
@@ -39,9 +39,9 @@ KernelMemoryBarMenuItem::KernelMemoryBarMenuItem(system_info& systemInfo)
 	fLastSum = -1;
 	fGrenze1 = -1;
 	fGrenze2 = -1;
-	fPhysicalMemory = systemInfo.max_pages * B_PAGE_SIZE / 1024LL;
-	fCommittedMemory = systemInfo.used_pages * B_PAGE_SIZE / 1024LL;
-	fCachedMemory = systemInfo.cached_pages * B_PAGE_SIZE / 1024LL;
+	fPhysicalMemory = (int64)systemInfo.max_pages * B_PAGE_SIZE / 1024LL;
+	fCommittedMemory = (int64)systemInfo.used_pages * B_PAGE_SIZE / 1024LL;
+	fCachedMemory = (int64)systemInfo.cached_pages * B_PAGE_SIZE / 1024LL;
 }
